requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
cffi==1.17.1
click==8.1.8
cryptography==45.0.5
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.1.1
h11==0.16.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jwt==1.3.1
mako==1.3.10
MarkupSafe==2.1.5
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-box==7.2.0
python-dotenv==1.0.1
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
zipp==3.20.2
